slight refactrign of JWT files, added readme
